feat(InvitesManager): create invitesmanager
